tools/python/requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.3
